requirements.txt
streamlit
python-docx
pandas
reportlab
rapidfuzz
plotly
matplotlib
